+++ src/cli/index.ts
@@ -1,185 +1,174 @@
-import { Command, CommandArg, ToolController } from "../core";
-import { controllers } from "../controllers";
-import { consoleTerminal, formatTable } from "../core/utils";
-import * as fs from "fs";
-import * as path from "path";
+import {
+    Command,
+    CommandArg,
+    getArgVariants,
+    ToolController,
+} from "../core";
+import {controllers} from "../controllers";
+import {
+    consoleTerminal,
+    formatTable,
+} from "../core/utils";
+import {printUsage} from "./help";
 
-enum ParseState {
-    OptionOrArg,
-    OptionValue,
+function findOptionArg(command: Command, name: string): CommandArg | undefined {
+    if (name.startsWith("--")) {
+        name = name.substr(2).toLowerCase();
+        return command.args?.find(x => !x.isArg && x.name === name);
+    }
+    if (name.startsWith("-")) {
+        name = name.substr(1).toLowerCase();
+        return command.args?.find(x => !x.isArg && x.alias === name);
+    }
+    return undefined;
 }
 
-type CommandLine = {
-    args: string[];
-    options: { [name: string]: any };
-}
+class CommandLine {
+    args: { help?: boolean, [name: string]: any } = {};
+    controller: ToolController | undefined = undefined;
+    command: Command | undefined = undefined;
+    positional: CommandArg[] = [];
+    unresolved = new Map<string, CommandArg>();
+    pending: CommandArg | undefined = undefined;
 
-function parseCommandLine(programArgs: string[]): CommandLine {
-    let state = ParseState.OptionOrArg;
-    let option = "";
-    const args: string[] = [];
-    const options: { [name: string]: any } = {};
+    setArgValue(arg: CommandArg, value: any) {
+        const name = arg.name
+            .split("-")
+            .map((x, i) => i > 0 ? (x.substr(0, 1).toUpperCase() + x.substr(1)) : x)
+            .join("");
+        this.args[name] = value;
+    }
 
-    for (const arg of programArgs) {
-        const argOption: string | null =
-            arg.startsWith("--") ? arg.substr(2) : (arg.startsWith("-") ? arg : null);
-        switch (state) {
-            case ParseState.OptionOrArg:
-                if (option !== "") {
-                    options[option] = "";
-                }
-                if (argOption) {
-                    option = argOption;
-                    state = ParseState.OptionValue;
-                } else {
-                    option = "";
-                    args.push(arg);
-                }
-                break;
-            case ParseState.OptionValue:
-                options[option] = arg;
-                option = "";
-                state = ParseState.OptionOrArg;
-                break;
+    async resolveValue(arg: CommandArg, value: string | undefined) {
+        if (arg.type === "boolean" && value === undefined) {
+            value = "true";
         }
-    }
-    if (option !== "") {
-        options[option] = "";
-    }
-    return {
-        args,
-        options,
-    };
-}
-
-export function printCommandUsage(controller: ToolController, command: Command) {
-    let usageArgs = "";
-    const options: CommandArg[] = [];
-    const args: CommandArg[] = [];
-    for (const arg of command.args ?? []) {
-        if (arg.isArg) {
-            usageArgs += ` ${arg.name}`;
-            args.push(arg);
+        if (value === undefined) {
+            throw new Error(`Missing value for ${arg.name}`);
+        }
+        let resolved: any;
+        if (arg.type === "boolean") {
+            resolved = value.toLowerCase() === "true";
         } else {
-            options.push(arg);
+            resolved = value;
+        }
+        const variants = await getArgVariants(arg);
+        if (variants && !variants.find(x => x.value === value)) {
+            throw await missingArgError(arg);
+        }
+        this.setArgValue(arg, resolved);
+        this.unresolved.delete(arg.name);
+        const i = this.positional.indexOf(arg);
+        if (i >= 0) {
+            this.positional.splice(i, 1);
+        }
+        if (this.pending === arg) {
+            this.pending = undefined;
         }
     }
-    if (options.length > 0) {
-        usageArgs += ` [options]`;
-    }
-    console.log(`Use: tondev ${controller.name} ${command.name}${usageArgs}`);
-    if (args.length > 0) {
-        console.log("Args:");
-        console.log(formatTable(args.map(x => ["  ", x.name, x.title])));
-    }
-    console.log("Options:");
-    console.log(`    --help, -h  Show command usage`);
-    console.log(formatTable(options.map(x => ["  ", `--${x.name}`, x.title])));
-    return;
-}
 
-export function printControllerUsage(controller: ToolController) {
-    const commands: [string, Command][] = controller.commands
-        .map(x => [`${controller.name} ${x.name}`, x]);
-    console.log(formatTable(commands.map(x => ["  ", x[0], x[1].title])));
-}
-
-export function printUsage(controller?: ToolController, command?: Command) {
-    const pkg = JSON.parse(fs.readFileSync(path.resolve(__dirname, "..", "..", "package.json"), "utf8"));
-    console.log(`TONDev Version: ${pkg.version}`);
-    if (controller && command) {
-        printCommandUsage(controller, command);
-        return;
-    }
-    console.log(`Use: tondev ${controller?.name ?? "tool"} ${command?.name ?? "command"} args [options]`);
-    console.log(`Options:`);
-    console.log(`    --help, -h  Show command usage`);
-    if (controller) {
-        console.log("Commands:");
-        printControllerUsage(controller);
-        return;
-    }
-    for (const controller of controllers) {
-        console.log(`\n${controller.title ?? controller.name} Commands:\n`);
-        printControllerUsage(controller);
-    }
-}
-
-function missingArgError(arg: CommandArg): Error {
-    const variants: string = arg.getVariants
-        ? "\n" +
-          formatTable(
-              [["Available variants:", ""], ...arg.getVariants().map(x => [x.name, x.description ?? ""])],
-              { headerSeparator: true }
-          )
-        : "";
-    throw new Error(`Missing required ${arg.name}${variants}`);
-}
-
-function getArgValue(arg: CommandArg, commandLine: CommandLine): string | undefined {
-    if (arg.isArg) {
-        const value = commandLine.args.splice(0, 1)[0];
-        if (value !== undefined) {
-            if (arg.getVariants === undefined) return value;
-            if (arg.getVariants().find(x => x.name === value)) return value;
-            throw missingArgError(arg);
-        }
+    async resolveDefault(arg: CommandArg) {
         if (arg.defaultValue !== undefined) {
-            return arg.defaultValue;
+            this.setArgValue(arg, arg.type === "boolean" ? arg.defaultValue === "true" : arg.defaultValue);
+        } else if (arg.type === "folder") {
+            this.setArgValue(arg, process.cwd());
+        } else {
+            throw await missingArgError(arg);
         }
-        throw missingArgError(arg);
     }
-    let value = commandLine.options[arg.name];
-    if (value !== undefined) {
-        return value;
+
+    async parseOptionName(name: string) {
+        if (this.pending) {
+            await this.resolveValue(this.pending, undefined);
+        }
+        const optionName: string = name.toLowerCase();
+        if (optionName === "--help" || optionName === "-h") {
+            this.args.help = true;
+        } else if (this.command) {
+            this.pending = findOptionArg(this.command, optionName);
+            if (!this.pending) {
+                throw new Error(`Unknown option ${optionName}`);
+            }
+            if (this.pending.type === "boolean") {
+                await this.resolveValue(this.pending, undefined);
+            }
+        } else if (this.controller) {
+            throw new Error(`Unexpected option ${optionName} before command name.`);
+        } else {
+            throw new Error(`Unexpected option ${optionName} before tool name.`);
+        }
     }
-    if (arg.defaultValue !== undefined) {
-        return arg.defaultValue;
+
+    async parse(programArgs: string[]) {
+        for (const arg of programArgs) {
+            if (arg.startsWith("-")) {
+                await this.parseOptionName(arg);
+            } else if (this.pending) {
+                await this.resolveValue(this.pending, arg);
+            } else if (this.controller && this.command) {
+                if (this.positional.length === 0) {
+                    throw new Error(`Unexpected argument ${arg}`);
+                }
+                await this.resolveValue(this.positional[0], arg);
+            } else if (this.controller) {
+                this.command = this.controller.commands.find(x => x.name === arg.toLowerCase());
+                if (!this.command) {
+                    throw new Error(`Unknown command: ${arg}`);
+                }
+                for (const arg of this.command.args ?? []) {
+                    this.unresolved.set(arg.name, arg);
+                    if (arg.isArg) {
+                        this.positional.push(arg);
+                    }
+                }
+            } else {
+                this.controller = controllers.find(x => x.name === arg.toLowerCase());
+                if (!this.controller) {
+                    throw new Error(`Unknown tool: ${arg}.`);
+                }
+            }
+        }
+        if (this.pending) {
+            await this.resolveValue(this.pending, undefined);
+        }
+        if (this.args.help) {
+            return;
+        }
+        for (const arg of this.unresolved.values()) {
+            await this.resolveDefault(arg);
+        }
     }
-    if (arg.type === "folder") {
-        return process.cwd();
-    }
-    throw missingArgError(arg);
 }
 
-function extractNextArg(commandLine: CommandLine): string {
-    return (commandLine.args.splice(0, 1)[0] ?? "").toLowerCase();
-}
-
-function getOption(commandLine: CommandLine, name: string, alias: string): string | undefined {
-    if (commandLine.options[name] !== undefined) {
-        return commandLine.options[name];
-    }
-    return commandLine.options[alias];
+async function missingArgError(arg: CommandArg): Promise<Error> {
+    const variants = await getArgVariants(arg);
+    const variantsString = variants
+        ? "\n" +
+        formatTable([
+                ["Available variants:", ""],
+                ...variants.map(x => [x.value, x.description ?? ""]),
+            ],
+            {headerSeparator: true},
+        )
+        : "";
+    throw new Error(`Missing required ${arg.name}${variantsString}`);
 }
 
 export async function run() {
-    const commandLine = parseCommandLine(process.argv.slice(2));
-    if (commandLine.args.length === 0) {
-        printUsage();
+    const parser = new CommandLine();
+    await parser.parse(process.argv.slice(2));
+    const {
+        controller,
+        command,
+        args,
+    } = parser;
+    if (!controller || !command) {
+        await printUsage(controller, command);
         return;
     }
-    const controllerName = extractNextArg(commandLine);
-    const controller = controllers.find(x => x.name === controllerName);
-    if (!controller) {
-        throw new Error(`Unknown tool: ${controllerName}.`);
-    }
-    const commandName = extractNextArg(commandLine);
-    const command = controller.commands.find(x => x.name === commandName);
-    if (!command) {
-        if (commandName) {
-            throw new Error(`Unknown command: ${commandName}`);
-        }
-        printUsage(controller, command);
+    if (parser.args.help) {
+        await printUsage(controller, command);
         return;
-    }
-    if (getOption(commandLine, "help", "-h") !== undefined) {
-        printUsage(controller, command);
-        return;
-    }
-    const args: { [name: string]: any } = {};
-    for (const arg of command.args ?? []) {
-        args[arg.name] = getArgValue(arg, commandLine);
     }
     await command.run(consoleTerminal, args);
 }